Create person class that other classes inherit from, create teacher_methods module for teachers, still a few tests not passing
