+++ student.rb
@@ -1,17 +1,6 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'person'
 
-  def initialize(options = {})
-    @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
-  end
-
+class Student < Person
   def set_phase(num)
     response = ""
     if num == @phase
